chore(config & fixeos): Configuraciones y fixeos de app - Modificaciones de los pom.xml - Instalacion Dependencia LiquiBase - Modificación de tipo valor (Integer) en CategoriaProducto y UnidadMedida - Modificaciones momentaneas de las propiedades de las apps

+++ ms_inventorario/src/main/java/ferrefix/ms_inventorario/model/CategoriaProducto.java
@@ -26,7 +26,7 @@ public class CategoriaProducto {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_categoria")
-    private Byte idCategoria;
+    private Integer idCategoria;
 
     @NotBlank(message = "El nombre de la categoría del producto es obligatorio")
     @Column(name = "nombre_categoria", nullable = false, length = 50, unique = true)
